+++ packages/swapkit/helpers/src/types/tokens.ts
@@ -1,7 +1,11 @@
 import type {
+  CaviarV1List,
   ChainflipList,
+  JupiterList,
   MayaList,
+  //   OciswapV1List,
   OneInchList,
+  OpenOceanV2List,
   PancakeswapList,
   PangolinList,
   SushiswapList,
@@ -14,9 +18,13 @@ import type {
 export type TokenTax = { buy: number; sell: number };
 
 export type TokenNames =
+  | (typeof CaviarV1List)["tokens"][number]["identifier"]
   | (typeof ChainflipList)["tokens"][number]["identifier"]
+  | (typeof JupiterList)["tokens"][number]["identifier"]
   | (typeof MayaList)["tokens"][number]["identifier"]
+  //   | (typeof OciswapV1List)["tokens"][number]["identifier"]
   | (typeof OneInchList)["tokens"][number]["identifier"]
+  | (typeof OpenOceanV2List)["tokens"][number]["identifier"]
   | (typeof PancakeswapList)["tokens"][number]["identifier"]
   | (typeof PangolinList)["tokens"][number]["identifier"]
   | (typeof SushiswapList)["tokens"][number]["identifier"]
@@ -24,9 +32,3 @@ export type TokenNames =
   | (typeof TraderjoeV2List)["tokens"][number]["identifier"]
   | (typeof UniswapV2List)["tokens"][number]["identifier"]
   | (typeof UniswapV3List)["tokens"][number]["identifier"];
-//   | (typeof CoinGeckoList)["tokens"][number]["identifier"]
-//   | (typeof PancakeswapETHList)["tokens"][number]["identifier"]
-//   | (typeof StargateARBList)["tokens"][number]["identifier"]
-//   | (typeof TraderjoeList)["tokens"][number]["identifier"]
-//   | (typeof WoofiList)["tokens"][number]["identifier"]
-//   | (typeof UniswapList)["tokens"][number]["identifier"];