+++ packages/plugins/evm/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @swapkit/plugin-evm
 
+## 1.2.0
+
+### Minor Changes
+
+- [#1094](https://github.com/thorswap/SwapKit/pull/1094) [`9421d52`](https://github.com/thorswap/SwapKit/commit/9421d52df6f0755d104b29d0367e1bf2c093e4e3) Thanks [@towanTG](https://github.com/towanTG)! - Adds new providers
+
+## 1.1.36
+
+### Patch Changes
+
+- Updated dependencies [[`a2f1571`](https://github.com/thorswap/SwapKit/commit/a2f157186da1ce4df60374371b2b5872119cd0b4), [`a2f1571`](https://github.com/thorswap/SwapKit/commit/a2f157186da1ce4df60374371b2b5872119cd0b4)]:
+  - @swapkit/helpers@1.16.0
+  - @swapkit/api@1.10.2
+  - @swapkit/toolbox-evm@1.7.2
+
 ## 1.1.35
 
 ### Patch Changes