+++ packages/wallets/talisman/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @swapkit/wallet-talisman
 
+## 1.3.9
+
+### Patch Changes
+
+- Updated dependencies [[`a2f1571`](https://github.com/thorswap/SwapKit/commit/a2f157186da1ce4df60374371b2b5872119cd0b4), [`a2f1571`](https://github.com/thorswap/SwapKit/commit/a2f157186da1ce4df60374371b2b5872119cd0b4)]:
+  - @swapkit/helpers@1.16.0
+  - @swapkit/toolbox-evm@1.7.2
+  - @swapkit/toolbox-substrate@1.3.2
+
 ## 1.3.8
 
 ### Patch Changes